+++ packages/core/example/eth/get-master-wallet-balance.ts
@@ -1,0 +1,20 @@
+/// <reference path="../../src/typings/index.d.ts" />
+import { SDK, Env } from "../../src";
+import { EthWallet } from "../../src/eth/wallet";
+import "dotenv/config";
+
+async function main() {
+  const sdk = new SDK({
+    accessToken: process.env.ACCESS_TOKEN,
+    secret: process.env.SECRET,
+    url: process.env.URL,
+    env: Env.Local,
+  });
+
+  const wallet = await sdk.eth.wallets.getMasterWallet(
+    "73846a39bcd6124a205834c05daea5f5"
+  );
+  console.log(await wallet.getBalance());
+}
+
+main();