+++ packages/core/example/verify-wallet-address.ts
@@ -1,6 +1,6 @@
 /// <reference path="../src/typings/index.d.ts" />
 import { SDK, Env } from "../src";
-import { EthMasterWallet } from "../src/eth/wallet";
+import { EthWallet } from "../src/eth/wallet";
 import "dotenv/config";
 
 async function main() {
@@ -11,7 +11,7 @@ async function main() {
     env: Env.Local,
   });
 
-  const wallet: EthMasterWallet = await sdk.eth.wallets.getMasterWallet(
+  const wallet = await sdk.eth.wallets.getMasterWallet(
     "c69c0890824c1f2533d39ba7f146869e"
   );
   // valid case